+++ app/javascript/components/internal-alert/component.jsx
@@ -73,7 +73,9 @@ const Component = ({ title, items, severity, customIcon }) => {
     return (
       <>
         <div className={css.icon}>{customIcon || renderIcon()}</div>
-        {titleMessage}
+        <span className={css.message} data-testid="internal-alert-message">
+          {titleMessage}
+        </span>
       </>
     );
   };
